requirements.txt
python-telegram-bot
requests
lxml
cloudscraper
mega.py